+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Upgrades from major to major versions, such a change from version `5.6.0` to `6.0.0`, might require local configuration updates to ensure compatibility with your current scripts. Make sure you have the latest version of the SDK using `pip install -U cortex-client`.
 
-## [1.0.0] - 2019-07-05
+
+## [1.0.1] - 2019-07-09
 ### Added
-* 
+* Additional unit tests for connections, content, and environment config
+
+## [1.0.0] - 2019-07-08
+### Added
+* First port/extract from the deprecated cortex-client module.